renamed cavity decoder so that the button shows up on srf home
