+++ displays/cavity_display/frontend/fault_decoder_display.py
